CachedResource replication reconciler should retry on AlreadyExists
The reconciler now returns an error and let's the work item
to be requeued instead of trying to resolve the error in place.
The previous code was missing handling ObjectMeta. Letting it fail
and retry is cleaner and handles object creation in one place.

On-behalf-of: @SAP robert.vasek@sap.com
Signed-off-by: Robert Vasek <robert.vasek@clyso.com>